Move PostsByTagSearch from models/tag/search.rs to models/post/search.rs
